Send a mail asking to connect JetPack to a newly created WordCamp site, instead of directly doing it from cron.
In a multisite, Jetpack assumes that connection requests are triggered manually, and expects various objects like request, cookies etc to be present. So making the connection automatic from cron was throwing errors. Now we will send a mail, which will have a one click link that will be used to make the connection instead of a cron doing it directly.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/mu-plugins@7501 74240141-8908-4e6f-9713-ba540dce6ec7